Update dependencies
It's been a long time since this has run and a lot of dependencies have
been updated, including Restify.  Upgrading Restify several major
versions meant a few tweaks to the calls, but nothing major.

I also updated the app to run on `/web/` instead of the root, and
adjusted the relatively references accordingly.